+++ lib/esc_pos_printer.dart
@@ -5,7 +5,5 @@
  * Copyright (c) 2019-2020. All rights reserved.
  * See LICENSE for distribution and usage details.
  */
-library esc_pos_printer;
-
 export 'src/enums.dart';
 export 'src/network_printer.dart';